Refactored Extensions to a Util Class
