+++ app/src/main/java/de/othr/plantico/ui/homescreen/WateringAdapter.kt
@@ -13,9 +13,8 @@ import de.othr.plantico.R
 import de.othr.plantico.database.entities.OwnedPlant
 import de.othr.plantico.databinding.ViewWateringItemBinding
 import de.othr.plantico.ui.PlantActivity
-import java.text.SimpleDateFormat
-import java.time.LocalDateTime
-import java.time.ZoneOffset
+import de.othr.plantico.addDays
+import de.othr.plantico.toPlanticoString
 import java.util.*
 
 class WateringAdapter(context: Context) : ListAdapter<OwnedPlant, WateringAdapter.WateringViewHolder>(OwnedPlantsComparator()) {
@@ -93,27 +92,3 @@ class WateringAdapter(context: Context) : ListAdapter<OwnedPlant, WateringAdapte
 
 }
 
-fun Date.toPlanticoString():String{
-    val suffixes = arrayOf(
-        "th", "st", "nd", "rd", "th", "th", "th", "th",
-        "th", "th","th", "th", "th", "th", "th", "th", "th", "th", "th", "th",
-        "th", "st", "nd", "rd", "th", "th", "th", "th", "th", "th", "th", "st")
-    val c = Calendar.getInstance()
-    c.time = this
-    val day = c[Calendar.DAY_OF_MONTH]
-    val month_name = SimpleDateFormat("MMMM", Locale.ENGLISH).format(c.getTime())
-    return ""+day + suffixes[day] + " " + month_name
-
-}
-
-fun Date.addDays(days: Int):Date{
-    val c: Calendar = Calendar.getInstance()
-    c.time = this
-    c.add(Calendar.DATE, days)
-    return c.time
-}
-
-fun Date.nowUTC():Date{
-    return Date.from(LocalDateTime.now().toInstant(ZoneOffset.UTC))
-
-}
